server/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
attrs==26.1.0
av==17.0.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
deepgram-sdk==6.0.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
elevenlabs==2.40.0
email-validator==2.3.0
eval_type_backport==0.3.1
exa-py==2.9.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.0
fastar==0.9.0
flatbuffers==25.12.19
flupy==1.2.3
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.73.0
googlemaps==4.10.0
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
Jinja2==3.1.6
jiter==0.13.0
livekit==1.1.3
livekit-agents==1.5.1
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-deepgram==1.5.0
livekit-plugins-elevenlabs==1.5.0
livekit-plugins-openai==1.5.0
livekit-plugins-silero==1.5.1
livekit-protocol==1.1.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
nest-asyncio==1.6.0
numpy==2.4.3
onnxruntime==1.24.4
openai==2.29.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pgvector==0.3.6
pillow==12.1.1
pluggy==1.6.0
postgrest==2.28.3
prometheus_client==0.24.1
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pyparsing==3.3.2
pyroaring==1.0.4
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
realtime==2.28.3
requests==2.32.5
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
sentry-sdk==2.55.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.48
starlette==1.0.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
types-protobuf==6.32.1.20260221
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
vecs==0.4.5
watchfiles==1.1.1
websockets==15.0.1
wrapt==1.17.3
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
